chore: Remove unnecessary test
The lambda client test is not
necessary currently due to not
having a defined way to test SDKs.

This commit removes the test that
tests the SDK and slightly refactors
some code with the removal of this.

Also this removes the parameter for
passing in a SDK since it is no longer
needed.
